+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 * Embedded [Maputnik](https://github.com/maputnik/editor) style editor
 * Service Info page with viewer code snippets 
 
+#### Bug Fixes
+
+* Fix parsing of `extent` option in `generate` command
+
 <a name="0.7.0"></a>
 ## 0.7.0 (2017-03-12)